+++ unstoppable-ios-app/domains-manager-ios/Services/Push/Entities/PushConnectionRequest.swift
