+++ unstoppable-ios-app/domains-manager-ios/Services/Push/Entities/PushGroupChatDTO.swift
@@ -1,27 +1,11 @@
 //
-//  PushChat.swift
+//  PushGroupChatDTO.swift
 //  domains-manager-ios
 //
 //  Created by Oleg Kuplin on 26.05.2023.
 //
 
 import Foundation
-
-struct PushChat: Codable, Hashable {
-    let chatId: String
-    let did: String
-    let wallets: String
-    let profilePicture: String
-    let publicKey: String
-    let about: String
-    let name: String
-    let threadhash: String
-    let intent: String
-    let intentSentBy: String
-    let intentTimestamp: Date
-    let combinedDID: String
-    let groupInformation: PushGroupChatDTO?
-}
 
 struct PushGroupChatDTO: Hashable, Codable {
     let members: [PushGroupChatMember]
@@ -40,11 +24,4 @@ struct PushGroupChatDTO: Hashable, Codable {
     let scheduleAt: Date?
     let scheduleEnd: Date?
     let groupType: String
-}
-
-struct PushGroupChatMember: Hashable, Codable {
-    let wallet: String
-    let publicKey: String?
-    let isAdmin: Bool
-    let image: String
 }
